Se corrigen comentarios de PR, se agregan archivos al front de DocumentTable y DocumentUploadModal; de BaseModal se quitó el menú que tenian los encazabezados
