+++ back/src/main/java/com/proyectofinal/proyectofinal/controller/DocumentController.java
@@ -14,43 +14,25 @@ import java.util.List;
 @RequestMapping("documents")
 public class DocumentController {
 
-    private final DocumentService service;
+        private final DocumentService service;
 
-    public DocumentController(DocumentService service) {
-        this.service = service;
-    }
+        public DocumentController(DocumentService service) {
+                this.service = service;
+        }
 
-    @PostMapping("/upload")
-    public ResponseEntity<DocumentDTO> upload(@RequestParam("file") MultipartFile file) throws IOException {
-        Document document = service.saveFile(file);
+        @PostMapping("/upload")
+        public ResponseEntity<DocumentDTO> upload(@RequestParam("file") MultipartFile file) throws IOException {
+                Document document = service.saveFile(file);
+                return ResponseEntity.ok(new DocumentDTO(document));
+        }
 
-        DocumentDTO dto = DocumentDTO.builder()
-                .externalId(document.getExternalId())
-                .fileName(document.getFileName())
-                .createdAt(document.getCreatedAt() != null ? document.getCreatedAt().toString() : null)
-                .uploadedBy(document.getAppUser() != null ? document.getAppUser().getEmail() : "Unknown")
-                .build();
+        @GetMapping
+        public ResponseEntity<List<DocumentDTO>> getAllDocuments() {
+                List<DocumentDTO> documentDTOs = service.getAllDocuments()
+                                .stream()
+                                .map(DocumentDTO::new)
+                                .toList();
 
-        return ResponseEntity.ok(dto);
-    }
-
-    // Listar todos los documentos
-    @GetMapping
-    public ResponseEntity<List<DocumentDTO>> getAllDocuments() {
-        List<DocumentDTO> documentDTOs = service.getAllDocuments()
-                .stream()
-                .map(doc -> DocumentDTO.builder()
-                        .externalId(doc.getExternalId())
-                        .fileName(doc.getFileName())
-                        .createdAt(
-                                doc.getCreatedAt() != null
-                                        ? doc.getCreatedAt().toString()
-                                        : null)
-                        .uploadedBy(
-                                doc.getAppUser() != null ? doc.getAppUser().getEmail() : null)
-                        .build())
-                .toList();
-
-        return ResponseEntity.ok(documentDTOs);
-    }
+                return ResponseEntity.ok(documentDTOs);
+        }
 }
