+++ back/src/main/java/com/proyectofinal/proyectofinal/dto/DocumentDTO.java
@@ -1,5 +1,6 @@
 package com.proyectofinal.proyectofinal.dto;
 
+import com.proyectofinal.proyectofinal.model.Document;
 import lombok.*;
 
 @Getter
@@ -12,4 +13,11 @@ public class DocumentDTO {
     private String fileName;
     private String createdAt;
     private String uploadedBy;
+
+    public DocumentDTO(Document doc) {
+        this.externalId = doc.getExternalId();
+        this.fileName = doc.getFileName();
+        this.createdAt = doc.getCreatedAt() != null ? doc.getCreatedAt().toString() : null;
+        this.uploadedBy = doc.getAppUser() != null ? doc.getAppUser().getEmail() : null;
+    }
 }
