+++ front/src/components/base/BaseTable.tsx
@@ -1,16 +1,16 @@
-import React, { useEffect, useState } from 'react';
-import { Box, Paper } from '@mui/material';
-import { DataGrid, type GridColDef } from '@mui/x-data-grid';
-import './base-table.css';
-import BaseInput from './BaseInput';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { useForm } from 'react-hook-form';
-import z from 'zod';
+import React, { useEffect, useState } from "react";
+import { Box, Paper } from "@mui/material";
+import { DataGrid, type GridColDef } from "@mui/x-data-grid";
+import "./base-table.css";
+import BaseInput from "./BaseInput";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+import z from "zod";
 
 export interface Column<T> {
   field?: keyof T;
   label: string;
-  align?: 'left' | 'right' | 'center';
+  align?: "left" | "right" | "center";
   render?: (value: any, row: T) => React.ReactNode;
   sortable?: boolean;
   flex?: number;
@@ -33,7 +33,7 @@ interface SearchForm {
 
 export const searchSchema = z.object({
   value: z.string(),
-})
+});
 
 export function BaseTable<T>({
   columns,
@@ -48,24 +48,25 @@ export function BaseTable<T>({
   const { register, watch } = useForm<SearchForm>({
     resolver: zodResolver(searchSchema),
     defaultValues: {
-      value: ""
-    }
-  })
+      value: "",
+    },
+  });
 
   function getFilteredRows(originalArray: T[]) {
-    return originalArray.filter(row => searchFields.some(field => (row[field] as string).includes(searchValue)));
+    return originalArray.filter((row) =>
+      searchFields.some((field) => (row[field] as string).includes(searchValue))
+    );
   }
 
-  const searchValue = watch('value');
-  
+  const searchValue = watch("value");
+
   useEffect(() => {
     if (searchValue.length && searchFields.length) {
-      setFilteredRows(getFilteredRows(rows))
+      setFilteredRows(getFilteredRows(rows));
     } else {
       setFilteredRows(rows);
     }
-  }, [rows, searchValue])
-
+  }, [rows, searchValue]);
 
   const gridColumns: GridColDef[] = columns.map((col) => ({
     field: col.field as string,
@@ -78,29 +79,28 @@ export function BaseTable<T>({
     renderCell: (params) => {
       const value = params.row[col.field];
       return col.render ? (
-        <Box className="custom-cell">
-          {col.render(value, params.row)}
-        </Box>
+        <Box className="custom-cell">{col.render(value, params.row)}</Box>
       ) : (
-        String(value ?? '')
+        String(value ?? "")
       );
     },
   }));
 
   return (
     <Paper className="base-table-paper" elevation={0}>
-      {!!searchFields.length && 
+      {!!searchFields.length && (
         <form>
           <BaseInput
-            {...register('value')}
-            placeholder={searchPlaceholder ?? ""} />
-          
+            {...register("value")}
+            placeholder={searchPlaceholder ?? ""}
+          />
         </form>
-      }
+      )}
       <DataGrid
         className="base-table-data-grid"
         autoHeight={autoHeight}
         rows={filteredRows}
+        disableColumnMenu // Desactiva el menú de columnas
         columns={gridColumns}
         getRowId={getRowId ?? ((row) => (row as any).id ?? JSON.stringify(row))}
         pageSizeOptions={[10, 15, 25, 50]}
@@ -109,7 +109,7 @@ export function BaseTable<T>({
         }}
         disableRowSelectionOnClick
         localeText={{
-          noRowsLabel: 'No hay datos disponibles',
+          noRowsLabel: "No hay datos disponibles",
         }}
       />
     </Paper>
