+++ front/src/service/DocumentService.ts
@@ -4,20 +4,16 @@ import type { Document } from "../model/Document";
 export default class DocumentService {
   static basePath = "/documents";
 
-  static async getDocuments(): Promise<Document[]> {
-    const res = await httpClient.get<Document[]>(this.basePath);
-    return res.data;
+  static getDocuments() {
+    return httpClient.get<Document[]>(this.basePath);
   }
 
-  static async uploadDocument(file: File): Promise<Document> {
+  static uploadDocument(file: File) {
     const formData = new FormData();
     formData.append("file", file);
 
-    const res = await httpClient.post<Document>(
-      `${this.basePath}/upload`,
-      formData,
-      { headers: { "Content-Type": "multipart/form-data" } }
-    );
-    return res.data;
+    return httpClient.post<Document>(`${this.basePath}/upload`, formData, {
+      headers: { "Content-Type": "multipart/form-data" },
+    });
   }
 }
